Refactor MongoDB connection and data access: move to new mongo module, enhance logging,
