+++ utils/mongo/mongo_dal.py
@@ -1,10 +1,10 @@
 from gridfs import GridFS
-from utils.logger import Logger
-Logger = Logger.get_logger(service_name="DataManager")
+
 
 class MongoDAL:
-    def __init__(self, connection, collection_name="files_collection"):
+    def __init__(self, connection, collection_name, logger):
         self.connection = connection
+        self.logger = logger
         self.fs = GridFS(self.connection.db, collection=collection_name)
 
     def insert_file(self, file_path: str, filename: str, metadata: dict):
@@ -12,8 +12,8 @@ class MongoDAL:
             with open(file_path, "rb") as f:
                 file_data = f.read()
             file_id = self.fs.put(file_data, filename=filename, metadata=metadata)
-            Logger.info(f"File inserted with ID: {file_id}")
+            self.logger.info(f"File inserted with ID: {file_id}")
             return file_id
         except Exception as e:
-            Logger.error(f"Error inserting file: {e}")
+            self.logger.error(f"Error inserting file: {e}")
             return None
